Summary:
Summary:
Several of the concerns in the previous reviews have been addressed. Most
notably, the Mongoose web server has been replaced by the GNU's libmicrohttpd,
which is available as a shared library.

There is also a basic security feature which can be improved later: any
message that makes a change in KTorrent requires a verificacion key. One thing
that could make this system safer could be to randomize that key from time
to time.

One problem I could not crack is that adding a mangnet link does not work as
expected. The code that adds magnet links is the same as the callback from
the code that adds magnet links from the callback menus, but when my plugin
tries to add a magnet link, they just appear in the "Magnet Downloads" panel
and they there with zero peers. I have no idea why.

Differential Revision: https://phabricator.kde.org/D11409